# KAFKA-9628 Replace Produce request/response with automated protocol