--- a/clients/src/main/java/org/apache/kafka/common/protocol/CommonFields.java
+++ b/clients/src/main/java/org/apache/kafka/common/protocol/CommonFields.java
@@ -32,14 +32,4 @@ public class CommonFields {
                     "If the epoch provided by the client is larger than the current epoch known to the broker, then " +
                     "the UNKNOWN_LEADER_EPOCH error code will be returned. If the provided epoch is smaller, then " +
                     "the FENCED_LEADER_EPOCH error code will be returned.");
-
-    // Group APIs
-    public static final Field.Str GROUP_ID = new Field.Str("group_id", "The unique group identifier");
-
-    // Transactional APIs
-    public static final Field.Str TRANSACTIONAL_ID = new Field.Str("transactional_id", "The transactional id corresponding to the transaction.");
-    public static final Field.NullableStr NULLABLE_TRANSACTIONAL_ID = new Field.NullableStr("transactional_id",
-            "The transactional id or null if the producer is not transactional");
-    public static final Field.Int64 PRODUCER_ID = new Field.Int64("producer_id", "Current producer id in use by the transactional id.");
-    public static final Field.Int16 PRODUCER_EPOCH = new Field.Int16("producer_epoch", "Current epoch associated with the producer id.");
 }