--- a/clients/src/main/java/org/apache/kafka/common/requests/ProduceResponse.java
+++ b/clients/src/main/java/org/apache/kafka/common/requests/ProduceResponse.java
@@ -118,6 +118,7 @@ public class ProduceResponse extends AbstractResponse {
     /**
      * this method is used by testing only.
      * TODO: refactor the tests which are using this method and then remove this method from production code.
+     * https://issues.apache.org/jira/browse/KAFKA-10697
      */
     public Map<TopicPartition, PartitionResponse> responses() {
         return data.responses()