… clients/src/main/java/org/apache/kafka/common/requests/RequestUtils.java
… clients/src/main/resources/common/message/ProduceRequest.json
… clients/src/main/resources/common/message/ProduceResponse.json
… clients/src/test/java/org/apache/kafka/clients/NetworkClientTest.java
… clients/src/test/java/org/apache/kafka/clients/producer/internals/SenderTest.java
… clients/src/test/java/org/apache/kafka/clients/producer/internals/TransactionManagerTest.java
… clients/src/test/java/org/apache/kafka/common/message/MessageTest.java
… clients/src/test/java/org/apache/kafka/common/requests/ProduceRequestTest.java
… clients/src/test/java/org/apache/kafka/common/requests/ProduceResponseTest.java
… clients/src/test/java/org/apache/kafka/common/requests/RequestResponseTest.java
… clients/src/test/java/org/apache/kafka/common/requests/SchemaTestUtils.java
… core/src/main/scala/kafka/server/KafkaApis.scala
… core/src/test/scala/integration/kafka/api/AuthorizerIntegrationTest.scala
… core/src/test/scala/integration/kafka/network/DynamicConnectionQuotaTest.scala
… core/src/test/scala/unit/kafka/network/SocketServerTest.scala
… core/src/test/scala/unit/kafka/server/EdgeCaseRequestTest.scala
… core/src/test/scala/unit/kafka/server/KafkaApisTest.scala
… core/src/test/scala/unit/kafka/server/ProduceRequestTest.scala
… core/src/test/scala/unit/kafka/server/RequestQuotaTest.scala
… jmh-benchmarks/src/main/java/org/apache/kafka/jmh/producer/ProducerRequestBenchmark.java
… jmh-benchmarks/src/main/java/org/apache/kafka/jmh/producer/ProducerResponseBenchmark.java