@lbradstreet 2020-11-03
@chia7712 @hachikuji for the ProduceResponse handling, is this the overall broker side regression since you need both the construction and toStruct?
```
construction regression: 3.293 -> 580.099 ns/op
toStruct improvement: 825.889 -> 318.530 ns/op

overall response: 3.293+825.889 (old) = 829.182 vs 898.629 (new)
```

Could you please also provide an analysis of the garbage generation using `gc.alloc.rate.norm`?