@chia7712 2020-11-09
> Can we summarize the regression here for a real world workload?

@ijuma I have attached benchmark result to description. I will loop more benchmark later.