@hachikuji 2020-11-10
For what it's worth, I think we'll get back whatever we lose here by taking `Struct` out of the serialization path.