@hachikuji 2020-11-16
Posting allocation flame graphs from the producer before and after this patch:

![Screen Shot 2020-11-16 at 3 02 42 PM](https://user-images.githubusercontent.com/12502538/99318608-def92f80-281c-11eb-9dbd-9c75d2093fff.png)

![Screen Shot 2020-11-16 at 3 03 02 PM](https://user-images.githubusercontent.com/12502538/99318616-e15b8980-281c-11eb-9024-c6cbb6c22230.png)

So we succeeded in getting rid of the extra allocations in the network layer!