@chia7712 2020-11-17
@hachikuji @ijuma Thanks for all feedback. I'm going to do more tests :)