@ijuma 2020-11-17
Thanks @chia7712. Can you check that the @hachikuji's perf test doesn't result in significantly increased latency?