@chia7712 2020-11-17
> Can you check that the @hachikuji's perf test doesn't result in significantly increased latency?

I don't observer significantly increased latency on my local. The env is shown below.

- INTEL 10900K 10C20T 3.70 GHz ~ 5.30 GHz
- DDR4-2933 128 GB
- SSD
- Ubuntu 20.04
- OpenJDK 11.0.9.1

**trunk** (53026b799c609373696382c685b9080630128af2)
> 250000000 records sent, 1143578.577571 records/sec (279.19 MB/sec), 109.01 ms avg latency, 422.00 ms max latency, 107 ms 50th, 132 ms 95th, 146 ms 99th, 263 ms 99.9th.

**patch**
> 250000000 records sent, 1202958.315088 records/sec (293.69 MB/sec), 103.60 ms avg latency, 432.00 ms max latency, 101 ms 50th, 122 ms 95th, 146 ms 99th, 278 ms 99.9th.